add big star animation
