+++ app/(tabs)/_layout.tsx
@@ -1,12 +1,18 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Ionicons, MaterialCommunityIcons } from "@expo/vector-icons";
 import { Tabs } from "expo-router";
 import Colors from "@/constants/Colors";
 import { useColorScheme } from "@/components/useColorScheme";
 import { useClientOnlyValue } from "@/components/useClientOnlyValue";
+import { fetch } from "../slices/rocketsSlice";
+import { useDispatch } from "react-redux";
 
 export default function TabLayout() {
   const colorScheme = useColorScheme();
+  const dispatch = useDispatch();
+  useEffect(() => {
+    dispatch(fetch());
+  }, []);
 
   return (
     <Tabs
